+++ src/todomvc/fulcro_todomvc/server.clj
@@ -2,7 +2,7 @@
   (:require
     [clojure.core.async :as async]
     [com.fulcrologic.fulcro.algorithms.misc :as util]
-    [com.fulcrologic.fulcro.server.api-middleware :refer [not-found-handler wrap-api]]
+    [com.fulcrologic.fulcro.server.api-middleware :as fmw :refer [not-found-handler wrap-api]]
     [com.wsscode.pathom.connect :as pc]
     [com.wsscode.pathom.core :as p]
     [immutant.web :as web]
@@ -120,18 +120,21 @@
 ;; setup for a given connect system
 (def parser
   (p/parallel-parser
-    {::p/env     {::p/reader [p/map-reader
-                              pc/parallel-reader
-                              pc/open-ident-reader]}
+    {::p/env     {::p/reader                 [p/map-reader
+                                              pc/parallel-reader
+                                              pc/open-ident-reader]
+                  ::pc/mutation-join-globals [:tempids]}
      ::p/mutate  pc/mutate-async
      ::p/plugins [(pc/connect-plugin {::pc/register my-resolvers})
                   (p/post-process-parser-plugin p/elide-not-found)
                   p/error-handler-plugin]}))
 
 (def middleware (-> not-found-handler
-                  (wrap-resource "public")
                   (wrap-api {:uri    "/api"
                              :parser (fn [query] (async/<!! (parser {} query)))})
+                  (fmw/wrap-transit-params)
+                  (fmw/wrap-transit-response)
+                  (wrap-resource "public")
                   wrap-content-type
                   wrap-not-modified))